Skills section added on front end of profile page

+++ Nop.Plugin.Widgets.mobSocial/Content/mobSocial/js/controllers/skills/skill-controller.js
@@ -6,7 +6,7 @@
             SkillService.getUserSkills(userId,
                 function(response) {
                     if (response.Success) {
-                        $scope.userSkills = response.UserSkills;
+                        $scope.skills = response.Skills;
                     }
                 });
         }
@@ -15,33 +15,55 @@
             SkillService.getSkill(id,
                function (response) {
                    if (response.Success) {
-                       $scope.skill = response.UserSkill;
+                       $scope.skill = response.Skill;
                    }
                });
         }
 
-        $scope.postSkill = function (id) {
-            SkillService.postSkill(id,
+        $scope.postSkill = function (skill) {
+            SkillService.postSkill(skill,
                function (response) {
                    if (response.Success) {
-                       $scope.skill = response.UserSkill;
+                       skill = response.Skill;
+                       $scope.skill = null;
+
+                       //reorder skills
+                       $scope.skills.sort(function(s1, s2) {
+                           return s1.DisplayOrder > s2.DisplayOrder;
+                       });
                    }
                });
         }
 
         $scope.delete = function (id) {
+            if (!confirm("Are you sure you wish to delete this skill?"))
+                return;
             SkillService.delete(id,
                function (response) {
                    if (response.Success) {
-                       if ($scope.userSkills && $scope.skill) {
-                           for (var i = 0; i < $scope.userSkills.length; i++) {
-                               if ($scope.userSkills[i].Id == $scope.skill.Id) {
-                                   $scope.userSkills.slice(i, 1);
+                       if ($scope.skills && $scope.skill) {
+                           for (var i = 0; i < $scope.skills.length; i++) {
+                               if ($scope.skills[i].Id == $scope.skill.Id) {
+                                   $scope.skills.slice(i, 1);
                                }
                            }
                        }
                    }
                });
+        }
+
+        $scope.edit = function (skill) {
+            $scope.skills = $scope.skills || [];
+            if (!skill) {
+                skill = { Id: 0 };
+                $scope.skills.push(skill);
+            }
+            $scope.skill = skill;
+        }
+        $scope.cancelEdit = function (skill) {
+            if (skill.Id == 0)
+                $scope.skills.pop();
+            $scope.skill = null;
         }
     }
 ]);
